% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_symb.R
\name{cr_symb}
\alias{cr_symb}
\title{A funcion for creating symbols}
\usage{
cr_symb(k, m)
}
\description{
This function obtains the m-surroundings by selecting the m-1 nearest neighbors of each observation, allowing for a degree of overlap of s.
}
\examples{
# Obtain symbols for k=2 classes and m-surroundings of size 5
symb25 <- cr_symb(2,5)
symb25$p_symb #Permutations symbols
symb25$c_symb #Combinations-totals symbols
}
\keyword{symbols}
