% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_spq.R
\name{dgp_spq}
\alias{dgp_spq}
\title{A funcion to generate qualitative process with spatial structure}
\usage{
dgp_spq(x = x, p = p, listw = listw, rho = rho)
}
\arguments{
\item{x}{matrix of point coordinate}

\item{p}{un vector indicando el porcentaje de cada clase. Su longitud debe coincidir con el número de clases. Su suma debe ser 1.}

\item{listw}{A \code{listw} object created for example by
\code{\link[spdep]{nb2listw}} from \pkg{spatialreg} package; if
\code{\link[spdep]{nb2listw}} not given, set to
the same spatial weights as the \code{listw} argument. It can
also be a spatial weighting matrix of order \emph{(NxN)} instead of
a \code{listw} object. Default = \code{NULL}.}

\item{rho}{nivel de dependencia espacial (un valor entre -1 y 1)}
}
\value{
Devuelve un FACTOR codificado con los primeros números naturales
}
\description{
This function genera un proceso de datos cualitativos con estructura espacial
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
#
N <- 1000
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
listw <- spdep::nb2listw(knn2nb(knearneigh(cbind(cx,cy), k=4)))
p <- c(1/6,3/6,2/6)
rho = 0.5
xf <- dgp_spq(x = x, p = p, listw = listw, rho = rho)

}
\seealso{
\code{\link{q_symb}}, \code{\link{m_surr_no}}
}
\keyword{m_surround}
