% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc.test.R
\name{jc.test}
\alias{jc.test}
\title{A function to compute jointcount test for
binomial and multinomial (asymptotic and permutation
distributions)}
\usage{
jc.test(formula = NULL,
               data = NULL,
               fx = NULL,
               listw = NULL,
               na.action,
               zero.policy = NULL,
               distr = "asymptotic",
               alternative = "greater",
               control =list())
}
\arguments{
\item{formula}{a symbolic description of the factor(s).}

\item{data}{an (optional) data frame or a sf object
with points/multipolygons geometry containing the
variable(s) to be tested.}

\item{fx}{a factor  or a matrix
of factors in columns}

\item{listw}{A \code{listw} object created for example by
\code{\link[spdep]{nb2listw}} from \pkg{spatialreg} package;
if \code{\link[spdep]{nb2listw}} not given,
the spatial weights are built using the object given
in \code{listw} argument (usually an \code{sf}
object). Default = \code{NULL}.}

\item{na.action}{A function (default \code{options("na.action")}),
can also be \code{na.omit} or \code{na.exclude} with consequences
for residuals and fitted values. It may be necessary to set
\code{zero.policy} to \code{TRUE} because this subsetting may
create no-neighbour observations.}

\item{zero.policy}{Similar to the corresponding parameter of
\code{\link[spatialreg]{lagsarlm}} function in
\pkg{spatialreg} package.
If \code{TRUE} assign zero to the lagged value of zones without
neighbours. Default = \code{NULL}.}

\item{distr}{character. Distribution type "asymptotic" (default) or "mc".}

\item{control}{Optional argument. See Control Argument section.}
}
\value{
An object of the class \code{htest}
}
\description{
A function to compute Q test for spatial qualitative data
}
\details{
Aquí Antonio escribe una linda historia ....
}
\section{Control arguments}{

\describe{
\item{nsim}{number of permutations for get the Monte Carlo distribution.
Default = 999}
\item{seedinit}{seed to select the initial element to star
the algorithm to get compute the m-surroundings or to start
the simulations.}
}
}

\examples{
 ## Case 1
 ## Multinomial + Binomial using a sf multipolygon
rm(list = ls())
data("Spain")
f1 <- ~ Older65 + MenWoman
jc1 <- jc.test(formula = f1,
               data = spain.sf,
               distr = "mc",
               alternative = "greater",
               zero.policy = TRUE)
summary(jc1)
f2 <- ~ MenWoman + Coast
jc2 <- jc.test(formula = f2,
               data = spain.sf,
               distr = "mc",
               zero.policy = TRUE)
summary(jc2)

# Case 2:
## Multinomial using a sf multipoint
rm(list = ls())
data("FastFood")
f1 <- ~ Type
jc3 <- jc.test(formula = f1,
               data = FastFood.sf,
               distr = "asymptotic",
               control = list(knn = 6))
 summary(jc3)

# Examples function joincount.test
data(oldcol)
HICRIME <- cut(COL.OLD$CRIME, breaks=c(0,35,80), labels=c("low","high"))
names(HICRIME) <- rownames(COL.OLD)
jc4 <- jc.test(fx = HICRIME,
               listw = nb2listw(COL.nb,
               style="B"))
 jc5 <- jc.test(fx = HICRIME,
                listw = nb2listw(COL.nb, style="B"),
                distr = "'mc")
 HICRIME <- cut(COL.OLD$CRIME, breaks=c(0,35,80), labels=c("low","high"))
 names(HICRIME) <- rownames(COL.OLD)
 jc6 <- jc.test(fx = HICRIME,
                listw = nb2listw(COL.nb,
                                 style="B"))
}
\references{
\itemize{ CAMBIAR...
\item Cliff, A. D., Ord, J. K. 1981 \emph{Spatial processes}, Pion, pp. 19-20.
\item Upton, G., Fingleton, B. 1985 \emph{Spatial data analysis by example: point pattern and qualitative data}, Wiley, pp. 158–170.
}
}
\seealso{
\code{\link{print.summary.spqtest}},
\code{\link[spdep]{joincount.test}},
\code{\link[spdep]{joincount.multi}}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paezha@gmail.com} \cr
Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
}
}
\keyword{data,}
\keyword{dependence}
\keyword{qualitative}
\keyword{spatial}
