% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_srq_test.R
\name{local_srq_test}
\alias{local_srq_test}
\title{A function to calculate the local spatial runs tests (SRQ)}
\usage{
srq_test(formula = NULL, data = NULL, na.action, xf = NULL,
listw = listw, alternative = "less" , nsim = NULL, control = list())
}
\arguments{
\item{data}{san optional data frame or a sf object containing the variable to testing for.}

\item{xf}{vector de observaciones de la clase factor.}

\item{listw}{una lista de vecinos (tipo knn o nb) o una matrix W que indique el orden de cada $m_i-entorno$
(por ejemplo de inversa distancia). Para calcular el numero de rachas en cada m_i-entorno debe establecerse
un orden, por ejemplo del vecino más próximo al más lejano.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one
of "two.sided", "greater" or "less" (default).}

\item{nsim}{Number of permutations to optain confidence intervals by permutation (CI).
Default value is NULL to don`t get CI of number of runs and local.}
}
\value{
decir que devuelve
  \tabular{ll}{
    \code{SR} \tab numero total de rachas \cr
    \code{dnr} \tab distribución empírica del numero de rachas \cr
    \code{SRQ} \tab Test de homogeneidad. Negative sign indicates global homogeneity \cr
    \code{p.valueSRQ} \tab p-value of the SRglobal \cr
    \code{p.valueSRQB} \tab the pseudo p-value of the SRQ test\cr
    \code{SRGP} \tab vector con los 'nsim' valores de SRGlobal por remuestreo permutacional\cr
    \code{SRQlocal} \tab a data-frame with "runs","mean","std","z-value","p-value" \cr
    \code{SRLP} \tab matrix de orden n x nsim con los valores de SRLocal por remuestreo permutacional\cr
    }
}
\description{
This function calculates the local spatial runs test for each localization.
The SQRL local tests
}
\details{
El objeto listw puede ser de tres tipos:
  \tabular{ll}{
    \code{knn} \tab Un objeto tipo knn obtenida utilizando el criterio del vecino más próximo, obtenida por ejemplo con \code{\link{knearneigh}}\cr
    \code{nb} \tab Un objeto tipo nb obtenido con \code{\link{spdep::poly2nb}}.\cr
    \code{matrix} \tab Una matriz indicando el orden de vecindad de cada m_i-entorno. Por ejemplo basada en inversa de la distancia con cierto punto de corte.\cr
    }
}
\examples{

# Local SRQ test based on knn
rm(list = ls())
N <- 100
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
listw <- knearneigh(cbind(cx,cy), k=3)
p <- c(1/6,3/6,2/6)
rho <- 0.5
xf <- dgp_spq(x = x, p = p, listw = listw, rho = rho)
lsrq <- local_srq_test(xf = xf, listw = listw)
lsrq$SRQlocal
plot.lsrq(lsrq)

# Fastfood example. sf (points)
data("FastFood")
x <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
listw <- spdep::knearneigh(x, k=4)
formula <- ~ Type
lsrq <- local_srq_test(formula = formula, data = FastFood.sf, listw = listw)
lsrq$SRQlocal
plot.srq(lsrq)

# With a sf object (poligons)
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
listw <- spdep::poly2nb(as(nc,"Spatial"), queen = FALSE)
p <- c(1/6,3/6,2/6)
rho = 0.5
co <- sf::st_coordinates(st_centroid(nc))
nc$xf <- dgp_spq(x = co, p = p, listw = listw, rho = rho)
plot(nc["xf"])
formula <- ~ xf
lsrq <- local_srq_test(formula = formula, data = nc, listw = listw, nsim=399)
lsrq$SRQlocal
plot.srq(srq)

# Con zonas aisladas sin conexion
data(Spain)
listw <- spdep::poly2nb(as(spain.sf,"Spatial"), queen = FALSE)
plot(spain.sf["HOM_MUJ"])
formula <- ~ HOM_MUJ
lsrq <- local_srq_test(formula = formula, data = spain.sf, listw = listw)
lsrq$SRQlocal

# SRQ test based on inverse distance
data("FastFood")
n = dim(FastFood.sf)[1]
dis <- 1000000/matrix(as.numeric(st_distance(FastFood.sf,FastFood.sf)),ncol=n,nrow=n)
diag(dis) <- 0
dis <- (dis < quantile(dis,.01))*dis

}
\seealso{
\code{\link{dgp_spq}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paez@gmail.com} \cr
  Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
  }
  @references
  \itemize{
    \item Ruiz, M., López, F., and Páez, A. (2010).
    A test for global and local homogeneity of categorical data based on spatial runs.
      \emph{Geographical Analysis}.
  }
}
\keyword{association,}
\keyword{qualitative}
\keyword{runs}
\keyword{spatial}
\keyword{test}
\keyword{variable,}
