% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_cbl.R
\name{m_surr_cbl}
\alias{m_surr_cbl}
\title{A funcion for creating m-surroudings from polygonal
       features maximizing the common borderline}
\usage{
m_surr_cbl(x = x, m = m, control = NULL)
}
\arguments{
\item{x}{input sf object with polygons or multipolygons features}

\item{m}{length of m-surrounding}

\item{s}{maximum overlap between any two m-surroundings}

\item{control}{Argumento opcional. Por definir}
}
\value{
REPASAR A list with four matrices: an m-surrounding matrix, named mh,
        whose rows correspond to the m-surroundings; a matrix, named mhs,
        containing the m-surroundings with maximum overlapping of s; a matrix including
        the length of borderlines between m-surroundings and a distance matrix,
        named dtmh, whose rows correspond to distance between the elements
        of m-surroundings.
}
\description{
This function obtains the m-surroundings by
             choosing the m-1 neighboring polygons. If there are more than
             (m-1) first order neighbors the algorithm chooses the polygons with the
             longest borderlines.
             **VIP** If there are less than (m-1) first order neighbors the
             corresponding m-surrounding is filled with second order
             neighbors in increasing order between centroid distances.
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
plot(sf::st_geometry(nc))
# VIP: THE THRESHOLD OF THE DISTANCE NEED TO BE VERY SMALL TO
# FIND M-SURROUNDINGS EXCEDING THE THRESHOLD.
lmh3 <- m_surr_cbl(nc, m = 3, s = 1, control = list(dthrpc = 0.05))
lmh5 <- m_surr_cbl(nc, m = 5, s = 2, control = list(dthrpc = 0.1))
dim(lmh3$mh); dim(lmh5$mh)
# m-surroundings or m-histories
 lmh3$mh[1:5, ]; lmh5$mh[1:5, ]
# length between borderlines in m-surroundings
lmh3$mhcbl[1:5, ]; lmh5$mhcbl[1:3, ]
# Distances in m-surroundings
lmh3$dtmh[1:5,]; lmh5$dtmh[1:5,]

}
\keyword{m_surrounding}
