% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_cdt.R
\name{m_surr_cdt}
\alias{m_surr_cdt}
\title{A funcion for creating m-surroudings minimizing the distances}
\usage{
m_surr_cdt(x = x, m = m, r = r, control = NULL)
}
\arguments{
\item{x}{input sf object with points/multipolygons geometry or matrix
of spatial coordinates}

\item{m}{length of m-surrounding}

\item{r}{maximum overlap between any two m-surroundings}

\item{control}{Argumento opcional. Por definir}
}
\value{
REPASAR A list with SOME matrices: an m-surrounding matrix, named ms,
       whose rows correspond to the m-surroundings and a distance matrix,
       named mdtms, whose rows correspond to distance between the elements
       of m-surroundings.....
}
\description{
This function obtains the m-surroundings by
             choosing the m-1 nearest centroids. If there are less
             than m-1 neighbors the corresponding m-surrounding
             is filled with NA's
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
N <- 1000
m = 5
r = 2
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
msurr_points <- m_surr_cdt(x = x, m = m, r = r,
                           control = list(dtmaxabs = 0.05))
ms <- msurr_points$ms
mdtms <- msurr_points$mdtms
dim(ms); dim(mdtms)
ms[1:10,]
mdtms[1:10,]
# Examples with multipolygons
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
plot(sf::st_geometry(nc))
lms3 <- m_surr_cdt(nc, m = 3, r = 2,
                   control = list(dtmaxpc = 0.1))
lms5 <- m_surr_cdt(nc, m = 5, r = 1,
                   control = list(dtmaxpc = 0.1))
# m-surroundings or m-histories
lms3$ms[1:3, ]; lms5$ms[1:3, ]
# Distances in m-surroundings
lms3$mdtms[1:3,]; lms5$mdtms[1:3,]
}
\keyword{m_surrounding}
