% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_maxs.R
\name{m_surr_maxs}
\alias{m_surr_maxs}
\title{A funcion to avoid overlaps of order greater than s in m-surroundings}
\usage{
m_surr_maxs(x = x, s = s, control = NULL)
}
\arguments{
\item{x}{matrix of m-surroundings}

\item{s}{maximum overlap between any two m-surroundings}

\item{control}{Argumento opcional. Por definir
VIP: ¿SE PUEDE PONER UNA DISTANCIA MÁXIMA PARA ELIMINAR
SURROUDINGS QUE SUPEREN ESE UMBRAL?.}
}
\value{
matrix of m-surroundings excluding cases with more than s overlaps
}
\description{
This function prunes the m-surroundings .
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
lmh3 <- m_surr_cbl(nc, m = 3)
mh3s <- m_surr_maxs(lmh3$mh, s = 1)
head(lmh3$mh); head(mh3s)
dim(lmh3$mh); dim(mh3s)
}
\keyword{m_surround}
