% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_no2.R
\name{m_surr_no2}
\alias{m_surr_no2}
\title{A funcion for creating m-surroudings ALTERNATIVA}
\usage{
m_surr_no2(x, m, s)
}
\arguments{
\item{x}{son coordenadas}

\item{m}{amplitud m-historia}

\item{s}{solapamiento máximo entre cualesquiera dos m-historias}
}
\description{
Esta función calcula m-historias. El algoritmo basado en m_surr_no2 obtiene m-historias
cuyos elementos pueden estar muy distantes (esto distorsiona el resultado del contraste). Este NUEVO algoritmo está basado en los knn.
Parte de una m-historia semilla (la correspondiente a la observación que se encuentra en primer lugar). La siguiente m-historia se añade siempre
que no solape en más de "s" elementos a todas las anteriores. El proceso se repite, incorporando nuevas m-historias siempre que no solapen
en más de "s" elementos a todas las anteriores.
SE OBTIENE MENOR NÚMERO DE M-HISTORIAS QUE CON LA VERSION CLÁSICA
PERO SON MÁS COMPACTAS.
NO HAY UN NÚMERO PREFIJADO DE M-HISTORIAS. EL ALGORITMO DEPENDE DE LA SEMILLA Y DEL ORDEN EN QUE SE INCLUYAN LAS M-HISTORIAS.
}
\details{
This function obtains the m-surroundings by selecting the m-1 nearest neighbors of each observation, allowing for a degree of overlap of s.
}
\examples{

# Obtain m-surroundings of size 3 (m=5), with degree of overlap of two (s=2)
N <- 100
m = 5
s = 2
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
mh <- m_surr_no2(x, m, s)
W <- matrix(0,ncol = N,nrow = N)
for (i in 1:dim(mh)[1]){
 W[mh[i,1],mh[i,2:m]]<-1
}
W <- mat2listw(W)
plot(W,x)



}
\keyword{m_surround}
