% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_no3.R
\name{m_surr_no3}
\alias{m_surr_no3}
\title{A funcion for creating m-surroudings (mejoradas en tiempo de cómputo)}
\usage{
m_surr_no_new(x = x, m = m, s = s, control = NULL)
}
\arguments{
\item{x}{input sf object with points/multipolygons geometry or matrix
of spatial coordinates}

\item{m}{amplitud m-historia}

\item{s}{solapamiento máximo entre cualesquiera dos m-surroundings}

\item{control}{Argumento opcional. Elimina aquellas m-surroundings cuyos elementos
no esten contenidos dentro del conjunto de los k=control vecinos más próximos.}
}
\value{
una matrix cuyas filas corresponden con las m-historias
}
\description{
This function obtains the m-surroundings by selecting the m-1 nearest neighbors of each observation, allowing for a degree of overlap of s.
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{

# Obtain m-surroundings of size 3 (m=5), with degree of overlap of two (s=2)
N <- 1000
m = 5
s = 2
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
system.time( msurr_points <- m_surr_no3(x = x, m = m, s = s,
                                        control = list(niter = 5,
                                        seedinit = 123456,
                                        dthr = 0.3)) )
mh <- msurr_points$mh
dtmh <- msurr_points$dtmh
dim(mh); dim(dtmh)
mh[1:10,]
dtmh[1:10,]


# Examples with multipolygons
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
plot(sf::st_geometry(nc))
system.time( msurr_polygonsf <- m_surr_no3(x = nc, m = 5, s = 2,
                     control = list(niter = 50, seedinit = 123456,
                     dthrpc = 0.20)) )
mh <- msurr_polygonsf$mh
dtmh <- msurr_polygonsf$dtmh
dim(mh); dim(dtmh)
mh[1:10,]
dtmh[1:10,]
}
\keyword{m_surround}
