% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_localsrq.R
\name{methods_localsrq}
\alias{methods_localsrq}
\alias{print.localsrq}
\alias{plot.localsrq}
\title{Methods for class localsrq}
\usage{
\method{print}{localsrq}(object, ...)

\method{plot}{localsrq}(object, sf = NULL, coor = NULL, sig = 0.05, viewplot = TRUE, ...)
}
\arguments{
\item{object}{a \code{localsrq} object created by \code{\link{qtest}}.}

\item{...}{further arguments passed to or from other methods.}

\item{viewplot}{logical value to show interactively the plots.
Default = \code{TRUE}}

\item{x}{similar to \code{object} argument for \code{print()}
and \code{plot} functions.}
}
\description{
Cambiar bla-bla por localsrq The \code{plot()} function allows the user to plot both beta and spatial
coefficients for all equations of the spsur model. The argument
\code{viewplot} is used to choose between interactive or non-interactive
plots. The \code{print()} function is used to print short tables including the values of beta and
spatial coefficients as well as p-values of significance test for each
coefficient.
\code{\link{summary.localsrq}} when a table (gt format) is needed.
The rest of methods works in the usual way.
}
\examples{
rm(list = ls()) # Clean memory
}
\references{
\itemize{
\item Ruiz, M., López, F., and Páez, A. (2010).
A test for global and local homogeneity of categorical data based on spatial runs.
\emph{Geographical Analysis}.
}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paezha@gmail.com} \cr
Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
}
}
