% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_spqdata.R
\name{methods_spqtest}
\alias{methods_spqtest}
\alias{plot.spqtest}
\title{Methods for class spqtest}
\usage{
\method{plot}{spqtest}(object, ci = 0.95, viewplot = TRUE, ...)
}
\arguments{
\item{object}{a \code{spqtest} object created by \code{\link{qtest}}.}

\item{ci}{confidence level for the intervals in \code{plot} method.
Default \code{ci = 0.95}}

\item{viewplot}{logical value to show interactively the plots.
Default = \code{TRUE}}

\item{...}{further arguments passed to or from other methods.}

\item{x}{similar to \code{object} argument for \code{print()}
and \code{plot} functions.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{lrtest}{logical value to compute likelihood ratio
test for nested models in \code{anova} method. Default = \code{TRUE}}
}
\description{
Cambiar bla-bla por spqtest The \code{plot()} function allows the user to plot both beta and spatial
coefficients for all equations of the spsur model. The argument
\code{viewplot} is used to choose between interactive or non-interactive
plots. The \code{print()} function is used to print short tables including the values of beta and
spatial coefficients as well as p-values of significance test for each
coefficient.
\code{\link{summary.spqtest}} when a table (gt format) is needed.
The rest of methods works in the usual way.
}
\examples{
rm(list = ls()) # Clean memory
}
\references{
\itemize{
\item Ruiz, M., López, F., and Páez, A. (2010).
A test for global and local homogeneity of categorical data based on spatial runs.
\emph{Geographical Analysis}.
}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paezha@gmail.com} \cr
Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
}
}
