% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb2nb_order.R
\name{nb2nb_order}
\alias{nb2nb_order}
\title{Una función para ordenar elementos de las m_i-historias}
\usage{
nb2nb_order <- function(listw = listw, sf = sf)
}
\arguments{
\item{listw}{una lista de vecinos tipo nb.}

\item{sf}{el objeto sf utilizado para .}
}
\value{
devuelve un objeto de la clase nb (elementos ordenados).
}
\description{
Es una función auxiliar. En el caso de obtener la lista de vecinos de clase nb con \code{\link{poly2nb}}
es necesario reordenar los elementos en función de la distancia y/o del angulo
}
\details{
Ordena los elementos de una lista nb. En primer lugar por distancia y en segundo lugar (a igualdad de distancias) por angulo.
}
\examples{

# With a sf object (irregular lattice)
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
listw <- spdep::poly2nb(as(nc,"Spatial"), queen = FALSE)
listw_order <- nb2nb_order(listw = listw, sf = nc)

# With a sf object (regular lattice: hexagons)
sfc = st_sfc(st_polygon(list(rbind(c(0,0), c(1,0), c(1,1), c(0,1), c(0,0)))))
hexs <- st_make_grid(sfc, cellsize = 0.1, square = FALSE)
hexs.sf <- st_sf(hexs)
listw  <- poly2nb(as(hexs.sf, "Spatial"), queen = FALSE)
listw_order <- nb2nb_order(listw = listw, sf = hexs.sf)

}
\seealso{
\code{\link{dgp_spq}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paez@gmail.com} \cr
  Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
  }
  @references
  \itemize{
    \item Ruiz, M., López, F., and Páez, A. (2010).
    A test for global and local homogeneity of categorical data based on spatial runs.
      \emph{Geographical Analysis}.
  }
}
\keyword{association,}
\keyword{qualitative}
\keyword{runs}
\keyword{spatial}
\keyword{test}
\keyword{variable,}
