% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lsrq.R
\name{plot.lsrq}
\alias{plot.lsrq}
\title{Plot the empirical distribution of runs.}
\usage{
\method{plot}{lsrq}(lsrq = lsrq, sf = NULL, coor = NULL, sig = 0.05)
}
\arguments{
\item{x}{object of class \emph{spq}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot the empirical distribution of runs.
}
\examples{
# Fastfood example. sf (points)
data("FastFood")
x <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
listw <- spdep::knearneigh(x, k = 2)
formula <- ~ Type
lsrq <- local_srq_test(formula = formula, data = FastFood.sf, listw = listw)
plot.lsrq(lsrq,sf=FastFood.sf)

# Spain example (poligons with 0 neinghbourhood)
data("Spain")
listw <- spdep::poly2nb(as(spain.sf,"Spatial"), queen = FALSE)
formula <- ~ HOM_MUJ
plot(spain.sf["HOM_MUJ"])
lsrq <- local_srq_test(formula = formula, data = spain.sf, listw = listw)
plot.lsrq(lsrq = lsrq, sf = spain.sf)

# With a sf object (poligons)
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
listw <- spdep::poly2nb(as(nc,"Spatial"), queen = FALSE)
p <- c(1/6,3/6,2/6)
rho = 0.5
co <- sf::st_coordinates(st_centroid(nc))
nc$xf <- dgp_spq(x = co, p = p, listw = listw, rho = rho)
plot(nc["xf"])
formula <- ~ xf
lsrq <- local_srq_test(formula = formula, data = nc, listw = listw)
lsrq$SRQlocal
plot.lsrq(lsrq,sf=nc)
}
\seealso{
\code{\link{srq_test}}.
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paez@gmail.com} \cr
  Manuel Ruiz  \tab \email{manuel.ruiz@upct.es} \cr
  }
}
