% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sprunstest.R
\name{plot.sprunstest}
\alias{plot.sprunstest}
\title{Plot the empirical distribution of runs}
\usage{
## S3 method for class 'sprunstest'
plot.sprunstest(srq = srq)
}
\arguments{
\item{srq}{A object of class \emph{sprunstest}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot the empirical distribution of runs
}
\details{
Plot the histogram with the empirical distribution
of the runs
}
\examples{
# Fastfood example. sf (points)
data("FastFood")
x <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
listw <- spdep::knearneigh(x, k = 2)
formula <- ~ Type
srq <- sp.runs.test(formula = formula, data = FastFood.sf, listw = listw, nsim = 299)
plot(srq)

# Spain example (poligons with 0 neinghbourhood)
data("Spain")
listw <- spdep::poly2nb(as(spain.sf,"Spatial"), queen = FALSE)
formula <- ~ Older65
srq <- sp.runs.test(formula = formula, data = spain.sf, listw = listw, nsim = 299)
plot(srq)
formula <- ~ MenWoman
srq <- sp.runs.test(formula = formula, data = spain.sf, listw = listw, nsim = 299)
plot(srq)
}
\seealso{
\code{\link{sp.runs.test}}.
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paez@gmail.com} \cr
Manuel Ruiz  \tab \email{manuel.ruiz@upct.es} \cr
}
}
