% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.srq.R
\name{plot.srq}
\alias{plot.srq}
\title{Plot the empirical distribution of runs.}
\usage{
\method{plot}{srq}(srq = srq)
}
\arguments{
\item{x}{object of class \emph{spq}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot the empirical distribution of runs.
}
\examples{
# Fastfood example. sf (points)
data("FastFood")
x <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
listw <- spdep::knearneigh(x, k = 2)
formula <- ~ Type
srq <- srq_test(formula = formula, data = FastFood.sf, listw = listw, nsim = 299)
plot.srq(srq)

# Spain example (poligons with 0 neinghbourhood)
data("Spain")
listw <- spdep::poly2nb(as(spain.sf,"Spatial"), queen = FALSE)
formula <- ~ MAY_65
srq <- srq_test(formula = formula, data = spain.sf, listw = listw, nsim = 299)
plot.srq(srq)
formula <- ~ HOM_MUJ
srq <- srq_test(formula = formula, data = spain.sf, listw = listw, nsim = 299)
plot.srq(srq)
}
\seealso{
\code{\link{srq_test}}.
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio Páez \tab \email{paez@gmail.com} \cr
  Manuel Ruiz  \tab \email{manuel.ruiz@upct.es} \cr
  }
}
