% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spjctest.R
\name{print.summary.spjctest}
\alias{print.summary.spjctest}
\title{Print method for objects of class summary.spjctest.}
\usage{
\method{print}{summary.spjctest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \emph{summary.spjctest}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for objects of class summary.spjctest.
}
\seealso{
\code{\link{summary.spqtest}}.
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
}
}
