% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_mc.R
\name{q_mc}
\alias{q_mc}
\title{A funcion to calculate Q using Monte Carlo}
\usage{
q_mc(Y, x, m, nsim = 999)
}
\arguments{
\item{x}{coordenadas asociadas a la observación de Y}

\item{m}{amplitud de las m-historias}

\item{nsim}{number of permutations}

\item{Y}{a factor of the same length as the coordinates x}

\item{s}{grado de solapamiento}
}
\value{
decir que cosas son las que devuelve
\tabular{ll}{
\code{Q0s_p} \tab value of the statistic of the observed distribution. Symbols p\cr
\code{p.value.p} \tab  the pseudo p-value of the Q0s_p test \cr
\code{efp_symb} \tab frecuencia empírica de los simbolos "p" de cada permutación\cr
\code{Q0s_c} \tab value of the statistic of the observed distribution. Symbols c\cr
\code{p.value.c} \tab  the pseudo p-value of the Q0s_c test \cr
\code{efc_symb} \tab frecuencia empírica de los simbolos "c" de cada permutación\cr
\code{nsim} \tab  nsim simulated values of statistic \cr
}
}
\description{
This function calculates Q0, a measure of spatial association based on symbolic entropy.
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
# Example 1
N <- 1000
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
listw <- spdep::nb2listw(spdep::knn2nb(spdep::knearneigh(cbind(cx,cy), k=4)))
p <- c(1/6,3/6,2/6)
rho = 0.5
fx <- dgp_spq(x = x, p = p, listw = listw, rho = rho)
Q0s <- q_mc(fx = fx, x = x, m = 3, nsim = 199)

# Example 2

# Load dataset
data("FastFood")
# Define coordinates
x <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
m <- 3
Q0s <- q_mc(fx = FastFood.sf$Type, x = x, m = 3, nsim = 199)
}
\seealso{
\code{\link{dgp_spq}}, \code{\link{m_surr_no}},\code{\link{q_symb}}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paezha@gmail.com} \cr
Manuel Ruiz  \tab \email{manuel.ruiz@upct.es} \cr
}
@references
\itemize{
\item Ruiz, M., López, F., and Páez, A. (2010).
Testing for spatial association of qualitative data using symbolic dynamics.
\emph{Journal of Geographical Systems}, 12(3), 281-309.
\item López, F., and Páez, A. (2012).
Distribution-free inference for Q(m) based on permutational bootstrapping: an application
to the spatial co-location pattern of firms in Madrid.
\emph{Estadística Española}, 177, 135-156.
}
}
\keyword{association,}
\keyword{entropy,}
\keyword{qualitative}
\keyword{spatial}
\keyword{symbolic}
\keyword{symbols}
\keyword{variable,}
