% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMap.R
\name{QMap}
\alias{QMap}
\title{A funcion to calculate QMap}
\usage{
QMap(df, mh, symb, type)
}
\arguments{
\item{df}{Es un data frame con tantas columnas como mapas a comparar}

\item{m}{longitud m-historia}

\item{type}{Type of symbols: "standard" or "compact". Default "standard"}

\item{r}{solapamiento}
}
\value{
decir que cosas son las que devuelve
  \tabular{ll}{
    \code{QMap} \tab valor del estadístico \cr
    \code{pmap_pval} \tab  p-value for QMap \cr
    \code{gl} \tab grados de libertad \cr
    \code{Tm} \tab  numero de mapas \cr
    \code{R} \tab  tamaño de cada mapa \cr
    }
}
\description{
This function COMPARA MAPAS based on symbolic entropy.
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
# Load dataset
N <- 1000
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
listw <- spdep::nb2listw(spdep::knn2nb(knearneigh(cbind(cx,cy), k=4)))
p <- c(2/6,2/6,2/6)
rho = 0.5
QY1 <- dgp_spq(x = x, p = p, listw = listw, rho = rho)
rho = 0.7
QY2 <- dgp_spq(x = x, p = p, listw = listw, rho = rho)
df = data.frame(cbind(QY1,QY2))
df$QY1 <- as.factor(df$QY1)
df$QY2 <- as.factor(df$QY2)
m=3
s=1
myqmap <- QMap(df, m=m, s=s, type ="standard")
myqmap$qmap
myqmap$qmap_pval
myqmap <- QMap(df, m=m, s=s, type ="compact")
myqmap$qmap
myqmap$qmap_pval
}
\seealso{
\code{\link{dgp_spq}}, \code{\link{m_surr_no}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio  \tab \email{paez@gmail.com} \cr
  Manolo  \tab \email{manuel.ruiz@upct.es} \cr
  }
  @references
  \itemize{
    \item Ruiz M, López FA and A Páez (2011).
    Comparison of Thematic Maps Using Symbolic Entropy.
      \emph{International Journal of Geographical Information Science},  26, 413-439.
    \item Ruiz, M., López, F., and Páez, A. (2010).
    Testing for spatial association of qualitative data using symbolic dynamics.
      \emph{Journal of Geographical Systems}, 12(3), 281-309.0.
  }
}
\keyword{association,}
\keyword{entropy,}
\keyword{qualitative}
\keyword{spatial}
\keyword{symbolic}
\keyword{symbols}
\keyword{variable,}
