% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.test.R
\name{scan.test}
\alias{scan.test}
\title{Compute the scan test.}
\usage{
scan.test(formula = NULL, data = NULL, fx = NULL, coor = NULL, case = NULL, nv = NULL,
nsim = NULL, distr = NULL, windows = "circular", alternative = "High", control = list())
}
\arguments{
\item{formula}{a symbolic description of the factor (optional).}

\item{data}{an (optional) data frame or a sf object containing the variable to testing for.}

\item{fx}{a factor (optional).}

\item{coor}{(optional) coordinates of observations.}

\item{case}{Only for bernoulli distribution. A element of factor, there are cases and non-cases for testing for cases versus non-cases}

\item{nv}{Maximum windows size, default nv = N/2. The algorithm scan for clusters of geographic size between 1
and the upper limit (nv) defined by the user.}

\item{nsim}{Number of permutations.}

\item{distr}{distribution of the spatial process: "bernoulli" for two levels or "multinomial" for three or more levels.}

\item{windows}{a string to select the type of cluster "circular" (default) of "elliptic".}

\item{alternative}{Only for bernoulli spatial process. A character string specifying the type of cluster, must be one
of "High" (default), "Both" or "Low".}

\item{control}{List of additional control arguments.}
}
\value{
A object of the \emph{htest} and \emph{scantest} class
\tabular{ll}{
\code{method} \tab The type of test applied ().\cr
\code{fx} \tab Factor included as input to get the scan test.\cr
\code{MLC} \tab Observations included into the Most Likelihood Cluster (MLC).\cr
\code{statistic} \tab Value of the scan test (maximum Log-likelihood ratio). \cr
\code{N} \tab Total number of observations.\cr
\code{nn} \tab Windows used to get the cluster.\cr
\code{nv} \tab Maximum number of observations into the cluster.\cr
\code{data.name} \tab A character string giving the name of the factor.\cr
\code{coor} \tab coordinates.\cr
\code{alternative} \tab Only for bernoulli spatial process. A character string describing the alternative hypothesis select by the user.\cr
\code{p.value} \tab p-value of the scan test.\cr
\code{cases.expect} \tab Expected cases into the MLC.\cr
\code{cases.observ} \tab Observed cases into the MLC.\cr
\code{nsim} \tab Number of permutations.\cr
}
}
\description{
This function compute the scan test for Bernoulli and Multinomial categorical spatial process.
}
\details{
Two alternative sets of arguments can be included in this function to compute the scan test:

\tabular{ll}{
\code{Option 1} \tab A factor (fx) and coordinates (coor). \cr
\code{Option 2} \tab A sf object (data) and the formula to specify the factor.
The function consider the coordinates of the centroids of the elements of th sf object. \cr
}

The spatial scan statistics are widely used in epidemiology, criminology or ecology.
Their purpose is to analyze the spatial distribution of points or geographical
regions by testing the hypothesis of spatial randomness his distribution on the
basis of different distributions (e.g. Bernoulli, Poisson or Normal distributions).
The \code{scan.test} function obtain the scan statistic for the Bernoulli and Multinomial distribution.

To test independence in a spatial process, under the null, the type of windows is irrelevant but under the alternative the elliptic
windows can to identify with more precision the cluster.

For big data sets (N >>) the windows = "elliptic" can be so slowly
}
\section{Control arguments}{

\tabular{ll}{
\code{seedinit} \tab Numerical value for the seed (only for boot version). Default value seedinit=123 \cr
}
}

\examples{

# Case 1: scan test bernoulli
rm(list = ls())
data(Spain)
formula <- ~ MenWoman
scan <- scan.test(formula = formula, data = spain.sf, case="men", nsim = 99, distr = "bernoulli")
print(scan)
summary(scan)
plot(scan, sf = spain.sf)
scan <- scan.test(formula = formula, data = spain.sf, case="men", nv = 5, nsim = 99, distr = "bernoulli")
print(scan)
plot(scan, sf = spain.sf)
scan <- scan.test(formula = formula, data = spain.sf, case="men", nv = 15, nsim = 99, distr = "bernoulli", windows ="elliptic")
print(scan)
scan <- scan.test(formula = formula, data = spain.sf, case="men", nv = 15, nsim = 99, distr = "bernoulli", windows ="elliptic", alternative = "Low")
print(scan)
plot.scantest(scan, sf = spain.sf)

# Case 2: scan test multinomial
rm(list = ls())
data(Spain)
formula <- ~ Older65
scan <- scan.test(formula = formula, data = spain.sf, nsim = 99, distr = "multinomial")
print(scan)
plot(scan, sf = spain.sf)

# Case 3: scan test multinomial
rm(list = ls())
data(FastFood)
formula <- ~ Type
scan <- scan.test(formula = formula, data = FastFood.sf, nsim = 99, distr = "multinomial", windows="elliptic", nv = 100)
print(scan)
summary(scan)
plot.scantest(scan, sf = FastFood.sf)

# Case 4: DGP two categories
rm(list = ls())
N <- 500
cx <- runif(N)
cy <- runif(N)
listw <- knearneigh(cbind(cx,cy), k = 10)
p <- c(1/2,1/2)
rho <- 0.5
fx <- dgp.spq(p = p, listw = listw, rho = rho)
scan <- scan.test(fx = fx, nsim = 9, case = "1", nv = 20, coor = cbind(cx,cy), distr = "bernoulli",windows="elliptic")
print(scan)
plot.scantest(scan)

# Case 5: DGP three categories
rm(list = ls())
N <- 1000
cx <- runif(N)
cy <- runif(N)
listw <- knearneigh(cbind(cx,cy), k = 10)
p <- c(1/3,1/3,1/3)
rho <- 0.5
fx <- dgp.spq(p = p, listw = listw, rho = rho)
scan <- scan.test(fx = fx, nsim = 19, coor = cbind(cx,cy), nv = 30, distr = "multinomial",windows="elliptic")
print(scan)
plot.scantest(scan)
}
\references{
\itemize{
\item Kulldorff M, Nagarwalla N. (1995).
Spatial disease clusters: Detection and Inference.
\emph{Statistics in Medicine}. 14:799-810
\item Jung I, Kulldorff M, Richard OJ (2010).
A spatial scan statistic for multinomial data.
\emph{Statistics in Medicine}. 29(18), 1910-1918
\item Páez, A., López-Hernández, F. A., Ortega-García, J. A., & Ruiz, M. (2016).
Clustering and co-occurrence of cancer types: A comparison of techniques with an application to pediatric cancer in Murcia, Spain.
\emph{Spatial Analysis in Health Geography}, 69-90.

}
}
\seealso{
\code{\link{local.sp.runs.test}}, \code{\link{dgp.spq}}, \code{\link{Q.test}},
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paezha@gmail.com} \cr
Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
}
}
\keyword{association,}
\keyword{clusters}
\keyword{qualitative}
\keyword{scan}
\keyword{spatial}
\keyword{test,}
\keyword{variable,}
