% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.test.R
\name{similarity.test}
\alias{similarity.test}
\title{Compute the similarity test.}
\usage{
similarity.test(formula = NULL, data = NULL, na.action, fx = NULL, listw = listw,
alternative = "two.sided", distr = "asymptotic", nsim = NULL, control = list())
}
\arguments{
\item{formula}{a symbolic description of the factor (optional).}

\item{data}{an (optional) data frame or a sf object containing the variable to testing for.}

\item{fx}{a factor (optional).}

\item{alternative}{a character string specifying the type of cluster, must be one
of "High" (default), "Both" or "Low".}

\item{distr}{A string. Distribution of the test "asymptotic" (default) or "bootstrap"}

\item{nsim}{Number of permutations.}

\item{control}{List of additional control arguments.}
}
\value{
A object of the \emph{htest}
\tabular{ll}{
\code{data.name} \tab a character string giving the names of the data.\cr
\code{statistic} \tab Value of the similarity test \cr
\code{N} \tab total number of observations.\cr
\code{Zmlc} \tab Elements in the Most Likelihood Cluster.\cr
\code{alternative} \tab a character string describing the alternative hypothesis. \cr
\code{p.value} \tab p-value of the similarity test \cr
\code{similiarity.mc} \tab values of the similarity test in each permutation. \cr
}
}
\description{
This function compute the similarity test for categorical spatial process.
}
\details{
Escribir aquí cosas
}
\section{Control arguments}{

\tabular{ll}{
\code{seedinit} \tab Numerical value for the seed (only for boot version). Default value seedinit=123 \cr
}
}

\examples{
rm(list = ls())
N <- 100
cx <- runif(N)
cy <- runif(N)
listw <- knearneigh(cbind(cx,cy), k = 3)
p <- c(1/4,1/4,1/4,1/4)
rho <- 0.5
fx <- dgp.spq(p = p, listw = listw, rho = rho)
W <- (nb2mat(knn2nb(listw)) >0)*1
similarity <- similarity.test(fx = fx, data = FastFood.sf, listw = listw)
print(similarity)

# Case 2: SRQ test with formula, a sf object (points) and knn
rm(list = ls())
data("FastFood")
coor <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
listw <- spdep::knearneigh(coor, k = 4)
formula <- ~ Type
similarity <- similarity.test(formula = formula, data = FastFood.sf, listw = listw)
print(similarity)

# Case 3:
rm(list = ls())
data("Spain")
listw <- spdep::poly2nb(as(spain.sf,"Spatial"), queen = FALSE)
formula <- ~ MenWoman
similarity <- similarity.test(formula = formula, data = spain.sf, listw = listw)
print(similarity)
}
\references{
\itemize{
\item Farber, S., Marin, M. R., & Paez, A. (2015).
Testing for spatial independence using similarity relations.
\emph{Geographical Analysis}. 47(2), 97-120.
}
}
\seealso{
\code{\link{sp.runs.test}}, \code{\link{dgp.spq}}, \code{\link{Q.test}}, , \code{\link{scan.test}}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
Antonio Páez \tab \email{paezha@gmail.com} \cr
Manuel Ruiz \tab \email{manuel.ruiz@upct.es} \cr
}
}
\keyword{association,}
\keyword{clusters}
\keyword{qualitative}
\keyword{scan}
\keyword{spatial}
\keyword{test,}
\keyword{variable,}
