% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spain.sf}
\alias{spain.sf}
\title{Spain by provinces}
\format{
a sf object with 50 rows and 7 variables:
\describe{
\item{Province}{Province name.}
\item{Population}{Population.}
\item{Older65}{Factor 3 categorias en funcion del porcen mayores.}
\item{AverageAge}{Ageed average.}
\item{MenWoman}{Factor 2 categorias men si porcen hombres mayor mujeres.}
\item{MassTransitSystems}{Has masive transport sistem.}
\item{Coast}{Has coast.}
}
}
\source{
Páez et al. (2020) \url{https://onlinelibrary.wiley.com/doi/full/10.1111/gean.12241}
}
\usage{
data(Spain)
}
\description{
Data set with areas with isolated areas
}
\references{
\itemize{
\item Paez, A., Lopez, F. A., Menezes, T., Cavalcanti, R., & Pitta, M. (2020). \emph{A Spatio‐Temporal Analysis of
the Environmental Correlates of COVID‐19 Incidence in Spain.}. Geographical Analysis.
}
}
\keyword{datasets}
