% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spqdata-package.R
\docType{package}
\name{spqdata-package}
\alias{spqdata}
\alias{spqdata-package}
\title{spqdata: A package for the exploratory analysis of cross-sectional spatial qualitative data}
\description{
A package for the exploratory analysis of qualitative data with georeferences.
}
\details{
Functionality included in \pkg{spqdata} are:
}
\section{Datasets}{
 \pkg{spqdata} includes two different datasets: spain and Fastfood. These sets are used to illustrate the capabilities of different functions. Briefly, their main characteristics are the following \cr
\itemize{
\item The \emph{FastFood} An object of class sf with georeferenced points of a selection of fast food restaurants in Toronto, Canada.
\item The \emph{spain.sp} An object of class sf with the boundaries of Spanish provinces and several socio-economic variables.
}
}

\references{
Breusch T, Pagan A (1980). The Lagrange multiplier test and its applications to model specification in econometrics. \emph{Review of Economic Studies} 47: 239-254.

LeSage, J., and Pace, R. K. (2009). \emph{Introduction to spatial econometrics}. Chapman and Hall/CRC.

López, F.A., Mur, J., and Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. \emph{Annals of Regional Science}, 53(1), 197-220.

López, F.A., Martínez-Ortiz, P.J., and Cegarra-Navarro, J.G. (2017). Spatial spillovers in public expenditure on a municipal level in Spain. \emph{Annals of Regional Science}, 58(1), 39-65.

Mur, J., López, F., and Herrera, M. (2010). Testing for spatial effects in seemingly unrelated regressions. \emph{Spatial Economic Analysis}, 5(4), 399-440.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/spqdata}
  \item Report bugs at \url{https://github.com/paezha/spqdata/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Román Minguez Salido \email{Roman.Minguez@uclm.es} (\href{https://orcid.org/0000-0002-0490-3181}{ORCID})
  \item Fernando López Hernandez \email{fernando.lopez@upct.es} (\href{https://orcid.org/0000-0002-5397-9748}{ORCID})
  \item Manuel Ruiz Marin \email{manuel.ruiz@upct.es} (\href{https://orcid.org/0000-0001-9228-6410}{ORCID})
}

}
\keyword{internal}
