% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spjctest.R
\name{summary.spjctest}
\alias{summary.spjctest}
\title{Summary of estimated objects of class \emph{spjctest}.}
\usage{
\method{summary}{spjctest}(object, ...)
}
\arguments{
\item{object}{An \emph{spjctest} object including a list of \emph{htest}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \emph{summary.spjctest}
}
\description{
This function summarizes estimated \emph{spjctest} objects.
The tables in the output include basic information for each test.
blablabla...
}
\examples{

}
\seealso{
\code{\link{print.summary.spqtest}},
\code{\link[spdep]{joincount.test}},
\code{\link[spdep]{joincount.multi}}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
}
}
