% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spqtest.R
\name{summary.spqtest}
\alias{summary.spqtest}
\title{Summary of estimated objects of class \emph{spqtest}.}
\usage{
\method{summary}{spqtest}(object, ...)
}
\arguments{
\item{object}{An \emph{spqtest} object including a list of \emph{htest}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \emph{summary.spqtest}
}
\description{
This function summarizes estimated \emph{spqtest} objects.
The tables in the output include basic information for each test.
blablabla...
}
\examples{

}
\seealso{
\code{\link{print.summary.spqtest}}
}
\author{
\tabular{ll}{
Fernando López  \tab \email{fernando.lopez@upct.es} \cr
Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
}
}
