% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FastFood.sf}
\alias{FastFood.sf}
\title{Distribution Fast-Food restaurants in Toronto}
\format{
A sf object with 877 rows and 4 variables:

\describe{
  \item{ID}{Identification}
  \item{Lat}{Latitude.}
  \item{Lon}{Longitude.}
  \item{Type}{Factor with 3 types of restaurant, "H" Hamburguer "P" Pizza "S" Shanwich.}
}
}
\source{
El papel con A y M en JGS
}
\usage{
data(FastFood)
}
\description{
Distribution Fast-Food restaurants in Toronto
}
\references{
\itemize{
    \item PRuiz M, López FA, A Páez. (2010). \emph{Testing for spatial association of qualitative
    data using symbolic dynamics}. Journal of Geographical Systems. 12 (3) 281-309
  }
}
\keyword{datasets}
