% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_cdt.R
\name{m_surr_cdt}
\alias{m_surr_cdt}
\title{A funcion for creating m-surroudings minimizing the distances}
\usage{
m_surr_cdt(x = x, m = m, control = NULL)
}
\arguments{
\item{x}{input sf object with points/multipolygons geometry or matrix
of spatial coordinates}

\item{m}{length of m-surrounding}

\item{s}{maximum overlap between any two m-surroundings}

\item{control}{Argumento opcional. Por definir}
}
\value{
REPASAR A list with SOME matrices: an m-surrounding matrix, named mh,
       whose rows correspond to the m-surroundings and a distance matrix,
       named dtmh, whose rows correspond to distance between the elements
       of m-surroundings.....
}
\description{
This function obtains the m-surroundings by
             choosing the m-1 nearest centroids. If there are less
             than m-1 neighbors the corresponding m-surrounding
             is filled with NA's
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{
N <- 1000
m = 5
s = 2
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
mh <- msurr_points$mh
dtmh <- msurr_points$dtmh
dim(mh); dim(dtmh)
mh[1:10,]
dtmh[1:10,]
# Examples with multipolygons
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
plot(sf::st_geometry(nc))
lmh3 <- m_surr_cdt(nc, m = 3, s = 2,
                   control = list(dthrpc = 0.1))
lmh5 <- m_surr_cdt(nc, m = 5, s = 1,
                   control = list(dthrpc = 0.1))
# m-surroundings or m-histories
lmh3$mh[1:3, ]; lmh5$mh[1:3, ]
# Distances in m-surroundings
lmh3$dtmh[1:3,]; lmh5$dtmh[1:3,]
}
\keyword{m_surrounding}
