% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_surr_no.R
\name{m_surr_no}
\alias{m_surr_no}
\title{A funcion for creating m-surroudings}
\usage{
m_surr_no(coords, m, s)
}
\description{
This function obtains the m-surroundings by selecting the m-1 nearest neighbors of each observation, allowing for a degree of overlap of s.
}
\examples{
# Load dataset
data("Newark")

# Use false origin UTM coordinates
coords <- cbind(Newark$X_UTM2,Newark$Y_UTM2)

# Obtain m-surroundings of size 3 (m=3), with degree of overlap of two (s=2)
mh32 <- m_surr_no(coords,3,2)

}
\keyword{m_surround}
