% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_symb.R
\name{q_symb}
\alias{q_symb}
\title{A funcion to calculate Q}
\usage{
q_symb(Y, mh, symb)
}
\arguments{
\item{Y, }{mh, symb}
}
\description{
This function calculates Q, a measure of spatial association based on symbolic entropy.
}
\examples{
# Load dataset
data("Newark")

# Use UTM coordinates
coords <- cbind(Newark$X_UTM2,Newark$Y_UTM2)

# Define variable for analysis. Three (k=3) ethnic groups
Y1 <- 1*Newark$NW + 2*Newark$IRISH + 3*Newark$GERMAN;#Ethnicity only

# Obtain symbols for 3 classes
symb35 <- cr_symb(3,5)

# Obtain m-surroundings of size 5 (m=5), with degree of overlap of one (s=1)
mh51 <- m_surr_no(coords,5,1)

results.35 <- q_symb(Y1,mh51,symb35)
results.35$qc #Qc for combinations-totals symbols
results.35$qc_pval #p-value for Qc
results.35$c_plot #plot bar chart of symbols with intervals of confidence
}
\keyword{association,}
\keyword{entropy,}
\keyword{qualitative}
\keyword{spatial}
\keyword{symbolic}
\keyword{symbols}
\keyword{variable,}
