% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_symb_A.R
\name{q_symb_A}
\alias{q_symb_A}
\title{A funcion to calculate Q CON PLOTS}
\usage{
q_symb(fx, mh, symb)
}
\arguments{
\item{fx}{vector de observaciones Factor o numérico?}

\item{mh}{m-historias}

\item{symb}{matrix con simbolos}
}
\value{
decir que cosas son las que devuelve
  \tabular{ll}{
    \code{pc} \tab Qp for combinations-totals symbols \cr
    \code{pc_pval} \tab  p-value for Qp \cr
    \code{qc} \tab Qc for combinations-totals symbols \cr
    \code{qc_pval} \tab  p-value for Qc \cr
    \code{p_symb} \tab  Matriz que lista los símbolos sin compactar \cr
    \code{efp_symb} \tab  Frecuancia absoluta de cada símbolo (no compactados)\cr
    \code{c_symb} \tab  Matriz que lista los símbolos COMPACTOS \cr
    \code{efc_symb} \tab  Frecuancia absoluta de cada símbolo (Compactado)\cr
    }
}
\description{
This function calculates Q, a measure of spatial association based on symbolic entropy.
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{

# Load dataset
data("FastFood")
# Define coordinates
x <- sf::st_coordinates(FastFood.sf)
#x <- cbind(FastFood.sf$Lat,FastFood.sf$Lon)
# Obtain m-surroundings of size 3 (m=3), with degree of overlap of one (s=1)
m <- 3
s <- 1
mh31 <- m_surr_no(x,m,s)
symb33 <- cr_symb(3,3)

# Plot the m-sourronding
N <- dim(x)[1]
W <- matrix(0,ncol = N,nrow = N)
for (i in 1:dim(mh31)[1]){W[mh31[i,1],mh31[i,2:m]] <- 1}
g1 = igraph::graph.adjacency(W)
mylist <- list()
for (i in 1:dim(mh31)[1]){mylist[[i]] <- mh31[i,]}





# Obtain the Q test
W <- spdep::mat2listw(W)
results.31 <- q_symb_A(FastFood.sf$Type,mh31,symb33)
}
\seealso{
\code{\link{dgp_spq}}, \code{\link{m_surr_no}}
}
\author{
\tabular{ll}{
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Román Mínguez  \tab \email{roman.minguez@uclm.es} \cr
  Antonio  \tab \email{paez@gmail.com} \cr
  Manolo  \tab \email{manuel.ruiz@upct.es} \cr
  }
  @references
  \itemize{
    \item Ruiz, M., López, F., and Páez, A. (2010).
    Testing for spatial association of qualitative data using symbolic dynamics.
      \emph{Journal of Geographical Systems}, 12(3), 281-309.
    \item Ruiz, M., López, F., and Páez, A. (2010).
    Testing for spatial association of qualitative data using symbolic dynamics.
      \emph{Journal of Geographical Systems}, 12(3), 281-309.0.
  }
}
\keyword{association,}
\keyword{entropy,}
\keyword{qualitative}
\keyword{spatial}
\keyword{symbolic}
\keyword{symbols}
\keyword{variable,}
