% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qasy.R
\name{qasy}
\alias{qasy}
\title{A funcion to compute q test for spatial qualitative data using
       asymptotic distribution}
\usage{
qasy(formula = NULL, data = NULL, na.action,
            m = 3, s = 1, type = "p", control = NULL,
            xf = NULL, coord = NULL)
}
\arguments{
\item{m}{length of m-surrounding}

\item{s}{solapamiento máximo entre cualesquiera dos m-surroundings}

\item{typems}{type of m-surrounding ("cbl","cdt","no"). Default = "no"}

\item{control}{Argumento opcional. Por definir}

\item{xf}{input sf object with points/multipolygons geometry or matrix
of spatial coordinates}

\item{type}{"p" or "c"}
}
\value{
Por definir, probablemente un objeto tipo spq que herede también
        de h-test, ANOVA test o similar...
}
\description{
This function ....
}
\details{
Aquí Antonio escribe una linda historia
}
\examples{

# Ejemplos Fernando
# Ejemplo 1
data("FastFood")
f1 <- ~ Type
lq <- qasy(formula = f1, data = FastFood.sf, m = 3, s = 1,
           typems = "no", control = list(seedinit = 1111))
lq
 # Ejemplo 2

data("Spain")
f1 <- ~ COSTA
lq <- qasy(formula = f1, data = spain.sf, m = 3, s = 1,
           typems = "no", control = list(seedinit = 1111))
lq
# Examples with multipolygons
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
qb79 <- quantile(nc$BIR79)
nc$QBIR79 <- (nc$BIR79 > qb79[2]) + (nc$BIR79 > qb79[3]) +
(nc$BIR79 >= qb79[4]) + 1
nc$QBIR79 <- as.factor(nc$QBIR79)
plot(nc["QBIR79"], pal = c("#FFFEDE","#FFDFA2", "#FFA93F", "#D5610D"),
     main = "BIR79 (Quartiles)")
sid79 <- quantile(nc$SID79)
nc$QSID79 <- (nc$SID79 > sid79[2]) + (nc$SID79 > sid79[3]) +
(nc$SID79 >= sid79[4]) + 1
nc$QSID79 <- as.factor(nc$QSID79)
plot(nc["QSID79"], pal = c("#FFFEDE","#FFDFA2", "#FFA93F", "#D5610D"),
     main = "SID79 (Quartiles)")
f1 <- ~ QSID79 + QBIR79
lq1nc <- qasy(formula = f1, data = nc, m = 5, s = 2,
           typems = "no", control = list(seedinit = 1111,
                                         dthrpc = 0.5) )
lq1nc$QSID79; lq1nc$QBIR79

lq2nc <- qasy(formula = f1, data = nc, m = 5, s = 2,
              typems = "cbl", control = list(dthrpc = 0.2) )
lq2nc$QSID79; lq2nc$QBIR79

lq3nc <- qasy(formula = f1, data = nc, m = 5, s = 2,
              typems = "cdt", control = list(dthrpc = 0.2) )
lq3nc$QSID79; lq3nc$QBIR79
# Examples with points and matrix of variables

xf <- matrix(c(nc$QBIR79, nc$QSID79), ncol = 2, byrow = TRUE)
mctr <- suppressWarnings(sf::st_centroid(sf::st_geometry(nc),
                                         of_largest_polygon = TRUE))
mcoor <- sf::st_coordinates(mctr)[,c("X","Y")]
lq <- qasy(xf = xf, mcoor = mcoor, m = 5, s = 2, typems = "no",
           control = list(seedinit = 1111,
                          dthrpc = 0.5))
lq


}
\keyword{...}
