% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spain.sf}
\alias{spain.sf}
\title{España por provincias}
\format{

}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/gean.12241}
}
\usage{
data(Spain)
}
\description{
Conjunto de datos de prueba. PARA PROBAR ELEMENTOS SIN VECINOS
}
\references{
\itemize{
    \item Paez, A., Lopez, F. A., Menezes, T., Cavalcanti, R., & Pitta, M. (2020). \emph{A Spatio‐Temporal Analysis of
     the Environmental Correlates of COVID‐19 Incidence in Spain.}. Geographical Analysis.
  }
}
\keyword{datasets}
