% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spqdata-package.R
\docType{package}
\name{spqdata-package}
\alias{spqdata}
\alias{spqdata-package}
\title{spqdata: A package for the exploratory analysis of spatial qualitative data}
\description{
A package for the exploratory analysis of qualitative data with georeferences.

This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample data set with their respective documentation.
}
\references{
\url{https://github.com/paezha/spqdata}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/spqdata}
  \item Report bugs at \url{https://github.com/paezha/spqdata/issues}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Román Minguez Salido \email{Roman.Minguez@uclm.es} (\href{https://orcid.org/0000-0002-0490-3181}{ORCID})
  \item Fernando López Hernandez \email{fernando.lopez@upct.es} (\href{https://orcid.org/0000-0002-5397-9748}{ORCID})
}


Román Mínguez Salido, , Economía Política y Hacienda Pública, Estadística Económica y Empresarial y Política Económica, Universidad de Castilla-La Mancha \email{Roman.Minguez@uclm.es}

Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\keyword{internal}
